needed_ports now saves if a rule is installed
Before was determing only via port state if a new
shared link forwarding rule needed to be installed,
now it is determined by it's own boolean. There
was a "race condition" between the moment the virtual
switch registered at the physical switch and the
port state changed that is now solved.