Forward packet out generated packet in
The latency between the controller and switch was
measured by sending packetouts to the switch that
contained an action to immediately send it back.
This caused the packetin to not have metadata set
since it didn't travel through the flowtable. Because
of this the hypervisor couldn't figure out where
to send the packetin. To fix this does the hypervisor
check if only 1 virtual switch depends on this
physical switch. In that case it had to have been
that switch so forward it to them.